use a single reconciliation api with multiples types support
