--- app/api/refine/reconcile/root.rb
@@ -1,13 +1,0 @@
-class Refine::Reconcile::Root < Grape::API
-  include Refine::Reconcile::Base
-
-  # before do
-  #   error!("401 Unauthorized", 401) unless authorized?
-  # end
-
-  mount Refine::Reconcile::Grades => '/grades'
-  mount Refine::Reconcile::Languages => '/languages'
-  mount Refine::Reconcile::Identities => '/identities'
-  mount Refine::Reconcile::ResourceTypes => '/resource_types'
-  mount Refine::Reconcile::Subjects => '/subjects'
-end
