+++ config/routes.rb
@@ -1,6 +1,6 @@
 Rails.application.routes.draw do
   mount API::V1::Root => '/api/v1'
-  mount Refine::Reconcile::Root => '/refine/reconcile'
+  mount Refine::Reconcile => '/refine/reconcile'
 
   devise_for :users
